Run the DAFi algorithm

Method to run general DAFi workflow. Will transform the data via FCSTrans and scale each channel to the range 0-4096 by default.

DAFi(rawfcs_path, config_path, reverse_path = NULL, k = 100, xdims = 10,
  ydims = 10, excludeGated = FALSE, method = "Kmeans",
  initializer = "kmeans++", plotCentroids = TRUE, writeOutput = TRUE,
  debugmode = FALSE)

Arguments

rawfcs_path
path to raw fcs file

config_path
path to DAFi cell gating hierarchy configuration file (inclusion gates)

reverse_path
path to DAFi cell gating hierarchy configuration file (exclusion gates)

k
specify number of clusters for Kmeans type algorithms (default 100)

xdims
number of SOM nodes in x direction for FlowSOM algorithm (default 10)

ydims
number of SOM nodes in y direction for FlowSOM algorithm (default 10)

excludeGated
*experimental* set TRUE to exclude previously gated channels in reclustering (default FALSE)

method
set clustering algorithm from the following choices: 'Kmeans' (default), 'fSOM'

initializer
set initializer for Kmeans methods from the following choices: 'optimal_init', 'quantile_init', 'kmeans++' (default), and 'random'; see ClusterR

plotCentroids
whether or not to plot diagnostic plots to see centroid locations

writeOutput
whether or not to generate output events and percentage tables

See also

cluster_run,scatterTransform,timeTransform

Examples

# Minimal parameter run (all defaults)
DAFi("SampleData.fcs","include.config")

#> Warning: cannot open file 'include.config': No such file or directory

#> Error in file(file, "rt"): cannot open the connection

# Or specify method and options for FlowSOM clustering
DAFi("SampleData.fcs","include.config","exclude.config", xdims=20, ydims=25, method='fSOM')

#> Warning: cannot open file 'include.config': No such file or directory

#> Error in file(file, "rt"): cannot open the connection

#' # Or specify method and options for Kmeans++ clustering
DAFi("SampleData.fcs","include.config","exclude.config", k=500, method='Kmeans', initializer='kmeans++')

#> Warning: cannot open file 'include.config': No such file or directory

#> Error in file(file, "rt"): cannot open the connection