Clustering algorithm wrapper function

Will generate the clustering results necessary for downstream filtering/gating and plotting

cluster_run(fcs_in, colsToUse, k = 100, xdims = 10, ydims = 10,
  method = "Kmeans", initializer = "kmeans++", debugmode = FALSE)

Arguments

fcs_in
compensated and transformed FCS in the flowFrame format as input

colsToUse
select the channels to use for the clustering: c(7,8,9,10)

k
select the number of desired clusters for the Kmeans-type clustering algorithms (default 100)

xdims
number of SOM nodes in x direction for FlowSOM algorithm (default 10)

ydims
number of SOM nodes in y direction for FlowSOM algorithm (default 10)

method
set clustering algorithm from the following choices: 'Kmeans' (default), 'fSOM'

initializer
set initializer for Kmeans methods from the following choices: 'optimal_init', 'quantile_init', 'kmeans++' (default), and 'random'; see ClusterR

Value

A list with four items: the first is the clustering result object directly from ClusterR or FlowSOM containing all information as originally constructed in the respective clustering packages, the second is the cluster membership mapping according the the original sequence from the input flowFrame, the third is the centroids in terms of the medianvalues of the columns from the flowFrame based on the membership mapping, and the fourth is the new flowFrame containing just the k (or xdims X ydims) number of centroids. This is a wrapper function for clustering methods included in ClusterR and FlowSOM Please see their documentations for more options.

See also

DAFi

Examples

# Kmeans with random initializer
clusterResult <- cluster_run(sampleflowFrame, colsToUse=c(9,12,14:18), k=500,
                      method="Kmeans", initializer="random")

#> Clustering via Kmeans 

#> Error in exprs(fcs_in): object 'sampleflowFrame' not found

# Kmeans with kmeans++ initializer
clusterResult <- cluster_run(sampleflowFrame, colsToUse=c(9,12,14:18), k=500)

#> Clustering via Kmeans 

#> Error in exprs(fcs_in): object 'sampleflowFrame' not found

# FlowSOM
clusterResult <- cluster_run(sampleflowFrame, colsToUse=c(9,12,14:18), xdims=20, ydims=25, method='fSOM')

#> Clustering via fSOM 

#> Error in ReadInput(input, pattern = pattern, compensate = compensate,     spillover = spillover, transform = transform, toTransform = toTransform,     transformFunction = transformFunction, scale = scale, scaled.center = scaled.center,     scaled.scale = scaled.scale, silent = silent): object 'sampleflowFrame' not found

#> Timing stopped at: 0.031 0.012 0.313